api/v1/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psutil==5.9.6
pydbus==0.6.0
requests==2.31.0

# gRPC dependencies para LND (OBRIGATÓRIO)
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4

# Wallet management dependencies
mnemonic==0.21
bip32==5.0.0
cryptography==41.0.7
base58==2.1.1
ecdsa==0.19.0
eth-hash[pycryptodome]==0.5.2
pycryptodome==3.20.0
